Implemented the VECTOR_DISTANCE function with the algorithms COSINE, EUCLIDEAN and DOT.
Syntax:
DISTANCE(vector, vector, [string])

Calculates the distance between two vectors per the specified calculation method.
It accepts the following arguments:
A column of VECTOR data type.
An input query of VECTOR data type.
A string that specifies the distance metric.
The supported values are "COSINE", "DOT", and "EUCLIDEAN".
If the distance metric is omitted the metric defaults to "DOT".

VECTOR_DISTANCE is a synonym for this function.